Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,15 +2,15 @@
 int main() {
     int a,b;char c;
     scanf("%d %d %c",&a,&b,&c);
-    int sum=(a+b);
-    int diff=(a-b);
-    int product=(a*b);
-    int div=(a/b);
+    float sum=(a+b);
+    float diff=(a-b);
+    float product=(a*b);
+    float div=(a/b);
     switch(c) {
-        case '+' : printf("%f",sum);break;
-        case '-' : printf("%f",diff);break;
-        case '*' : printf("%f",product);break;
-        case '/' : printf("%f",div);break;
+        case '+' : printf("%d",sum);break;
+        case '-' : printf("%d",diff);break;
+        case '*' : printf("%d",product);break;
+        case '/' : printf("%d",div);break;
         default : printf("error");break;
         }
         return  0;
